[Add: 리뷰 기능]

+++ src/main/java/cafeSubscription/coffee/domain/review/Review.java
@@ -23,7 +23,7 @@ import java.util.List;
 public class Review {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer reviewId;
+    private Long reviewId;
 
     @ManyToOne
     private Cafe cafe;
